+++ packages/core/src/components/widgets/SelectWidget.tsx
@@ -3,11 +3,11 @@ import { useCallback } from 'react';
 import type { FormContextType, RJSFSchema, StrictRJSFSchema, WidgetProps } from '@rjsf/utils';
 import {
   ariaDescribedByIds,
-  enumOptionsIndexForValue,
   enumOptionSelectedValue,
   enumOptionValueDecoder,
   enumOptionValueEncoder,
   getOptionValueFormat,
+  logUnsupportedDefaultForEnum,
 } from '@rjsf/utils';
 
 function getValue(event: SyntheticEvent<HTMLSelectElement>, multiple: boolean) {
@@ -71,16 +71,7 @@ function SelectWidget<T = any, S extends StrictRJSFSchema = RJSFSchema, F extend
 
   const selectValue = enumOptionSelectedValue<S>(value, enumOptions, multiple, optionValueFormat, emptyValue);
   const showPlaceholderOption = !multiple && schema.default === undefined;
-  if (
-    !multiple &&
-    Array.isArray(enumOptions) &&
-    schema.default !== undefined &&
-    enumOptionsIndexForValue<S>(schema.default, enumOptions, multiple) === undefined
-  ) {
-    console.error(
-      `The schema default value "${schema.default}" is not one of the values in the enum options for "${id}"`,
-    );
-  }
+  logUnsupportedDefaultForEnum<S>(id, schema, enumOptions, multiple);
 
   return (
     // oxlint-disable-next-line jsx-a11y/no-autofocus